Reconstruction/VideoVAEPlus/requirements.txt
av==12.0.0
accelerate==0.34.2 
academictorrents==2.3.3
albumentations==1.4.16
apex==0.9.10dev
beautifulsoup4==4.12.3
decord==0.6.0
diffusers==0.30.3
einops==0.8.0
fairscale==0.4.13
ftfy==6.2.3
huggingface-hub==0.23.2
imageio==2.33.1
kornia==0.7.3
moviepy==1.0.3
more_itertools==10.5.0
numpy==1.26.3
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.1.105
nvidia-nvtx-cu12==12.1.105
omegaconf==2.3.0
opencv-python==4.9.0.80
packaging==24.0
pandas==2.2.1
psutil==6.0.0
Pillow==10.4.0
pytorch_lightning==1.9.4
PyYAML==6.0.1
protobuf==3.20.*
Requests==2.32.3
safetensors==0.4.5
scipy==1.14.1
sentencepiece==0.2.0
tensorboard==2.18.0
taming-transformers==0.0.1
tensorboardX==2.6.2.2
timm==1.0.9
torch==2.2.0
torchaudio==2.2.0
torchmetrics==1.3.1
torchvision==0.17.0
tokenizers==0.13.3
tqdm==4.66.2
transformers==4.25.1
typing_extensions==4.12.2
xformers==0.0.24
lpips